# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VID Master Zimbabwe API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  city: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  learner: 'learner',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  city?: string;
  phone?: string;
  avatarUrl?: string;
  language: string;
  soundEnabled: number;
  role: UserRole;
  xp: number;
  coins: number;
  level: number;
  unlockedItems: string;
  streak: number;
  totalTests?: number;
  /** @nullable */
  passRate?: number | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type QuestionDifficulty = typeof QuestionDifficulty[keyof typeof QuestionDifficulty];


export const QuestionDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type QuestionStatus = typeof QuestionStatus[keyof typeof QuestionStatus];


export const QuestionStatus = {
  draft: 'draft',
  approved: 'approved',
  published: 'published',
  archived: 'archived',
} as const;

export interface Question {
  id: number;
  text: string;
  options: string[];
  correctAnswer: number;
  category: string;
  difficulty: QuestionDifficulty;
  explanation: string;
  /** @nullable */
  imageUrl?: string | null;
  status?: QuestionStatus;
  createdAt?: string;
}

export type QuestionInputDifficulty = typeof QuestionInputDifficulty[keyof typeof QuestionInputDifficulty];


export const QuestionInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface QuestionInput {
  text: string;
  /**
     * @minItems 3
     * @maxItems 4
     */
  options: string[];
  correctAnswer: number;
  category: string;
  difficulty: QuestionInputDifficulty;
  explanation: string;
  imageUrl?: string;
}

export type QuestionPatchDifficulty = typeof QuestionPatchDifficulty[keyof typeof QuestionPatchDifficulty];


export const QuestionPatchDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type QuestionPatchStatus = typeof QuestionPatchStatus[keyof typeof QuestionPatchStatus];


export const QuestionPatchStatus = {
  draft: 'draft',
  approved: 'approved',
  published: 'published',
  archived: 'archived',
} as const;

export interface QuestionPatch {
  text?: string;
  options?: string[];
  correctAnswer?: number;
  category?: string;
  difficulty?: QuestionPatchDifficulty;
  explanation?: string;
  imageUrl?: string;
  status?: QuestionPatchStatus;
}

export type StartTestInputMode = typeof StartTestInputMode[keyof typeof StartTestInputMode];


export const StartTestInputMode = {
  timed: 'timed',
  practice: 'practice',
} as const;

export interface StartTestInput {
  mode?: StartTestInputMode;
}

export interface AnswerInput {
  questionId: number;
  selectedAnswer: number;
}

export interface TestSubmission {
  answers: AnswerInput[];
}

export interface TestSession {
  sessionId: string;
  questions: Question[];
  startedAt: string;
  expiresAt: string;
  durationSeconds: number;
  mode?: string;
}

export interface AnswerDetail {
  questionId: number;
  text?: string;
  selectedAnswer: number;
  correctAnswer: number;
  isCorrect: boolean;
  explanation?: string;
}

export interface TestResult {
  sessionId: string;
  score: number;
  total: number;
  percentage: number;
  passed: boolean;
  timeTaken: number;
  completedAt: string;
  mode?: string;
  answers?: AnswerDetail[];
}

export interface UserStats {
  totalTests: number;
  totalQuestions: number;
  correctAnswers: number;
  accuracy: number;
  streak: number;
  xp: number;
  level: number;
  rank: string;
  examReadiness?: number;
  weeklyActivity?: number[];
}

export interface DashboardSummary {
  recentTests: TestResult[];
  accuracy: number;
  streak: number;
  examReadiness: number;
  totalTests: number;
  xp: number;
  level: number;
  rank: string;
  weeklyActivity?: number[];
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  name: string;
  city?: string;
  xp: number;
  level: number;
  accuracy: number;
  totalTests: number;
}

export interface CategoryStat {
  category: string;
  total: number;
  correct: number;
  accuracy: number;
}

export interface BookmarkInput {
  questionId: number;
}

export interface RoadSign {
  id: number;
  name: string;
  category: string;
  meaning: string;
  imageUrl: string;
  /** @nullable */
  usage?: string | null;
}

export interface RoadSignInput {
  name: string;
  category: string;
  meaning: string;
  imageUrl: string;
  usage?: string;
}

export interface AdminStats {
  totalUsers: number;
  totalQuestions: number;
  totalTests: number;
  passRate: number;
  activeToday: number;
}

export type UserRoleUpdateRole = typeof UserRoleUpdateRole[keyof typeof UserRoleUpdateRole];


export const UserRoleUpdateRole = {
  learner: 'learner',
  admin: 'admin',
} as const;

export interface UserRoleUpdate {
  role: UserRoleUpdateRole;
}

export type ListQuestionsParams = {
category?: string;
difficulty?: string;
limit?: number;
offset?: number;
};

export type GetTestHistoryParams = {
limit?: number;
offset?: number;
};

export type GetLeaderboardParams = {
period?: GetLeaderboardPeriod;
};

export type GetLeaderboardPeriod = typeof GetLeaderboardPeriod[keyof typeof GetLeaderboardPeriod];


export const GetLeaderboardPeriod = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
  alltime: 'alltime',
} as const;

export type ListSignsParams = {
category?: string;
search?: string;
};

export type ListUsersParams = {
limit?: number;
offset?: number;
};