# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VID Master Zimbabwe API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions (with optional filters)
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: difficulty
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a new question (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get question by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # --- TESTS ---
...
  /signs:
    get:
      operationId: listSigns
      tags: [signs]
      summary: List road signs (with optional category filter)
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Road signs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoadSign"
    post:
      operationId: createSign
      tags: [signs]
      summary: Create a road sign (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoadSignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadSign"
  /signs/{id}:
    get:
      operationId: getSign
      tags: [signs]
      summary: Get road sign by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Road sign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadSign"
    patch:
      operationId: updateSign
      tags: [signs]
      summary: Update a road sign (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoadSignInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadSign"
    delete:
      operationId: deleteSign
      tags: [signs]
      summary: Delete a road sign (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # --- ADMIN ---
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Update a user's role (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth
...
    User:
      type: object
      required: [id, email, name, role, xp, level, streak, createdAt, language, soundEnabled]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        city:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        language:
          type: string
        soundEnabled:
          type: integer
        role:
          type: string
          enum: [learner, admin]
        xp:
          type: integer
        coins:
          type: integer
        level:
          type: integer
        unlockedItems:
          type: string
        streak:
          type: integer
        totalTests:
          type: integer
        passRate:
          type: ["number", "null"]
        createdAt:
          type: string
...
    Question:
      type: object
      required: [id, text, options, correctAnswer, category, difficulty, explanation]
      properties:
        id:
          type: integer
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: integer
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        explanation:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, approved, published, archived]
        createdAt:
          type: string
    QuestionInput:
      type: object
      required: [text, options, correctAnswer, category, difficulty, explanation]
      properties:
        text:
          type: string
        options:
          type: array
          items:
            type: string
          minItems: 3
          maxItems: 4
        correctAnswer:
          type: integer
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        explanation:
          type: string
        imageUrl:
          type: string
    QuestionPatch:
      type: object
      properties:
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: integer
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        explanation:
          type: string
        imageUrl:
          type: string
        status:
          type: string
          enum: [draft, approved, published, archived]

    # Tests
...
    RoadSign:
      type: object
      required: [id, name, category, meaning, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        meaning:
          type: string
        imageUrl:
          type: string
        usage:
          type: ["string", "null"]
    RoadSignInput:
      type: object
      required: [name, category, meaning, imageUrl]
      properties:
        name:
          type: string
        category:
          type: string
        meaning:
          type: string
        imageUrl:
          type: string
        usage:
          type: string

    # Admin
...
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [learner, admin]